+++ octopoes/bits/runner.py
@@ -1,9 +1,14 @@
+import json
 from collections.abc import Iterator
+from datetime import datetime
 from importlib import import_module
 from inspect import isfunction, signature
 from typing import Any, Protocol
 
+from pydantic import JsonValue, TypeAdapter
+
 from bits.definitions import BitDefinition
+from octopoes.config.settings import CACHE_BITS
 from octopoes.models import OOI
 
 
@@ -18,8 +23,21 @@ class Runnable(Protocol):
 class BitRunner:
     def __init__(self, bit_definition: BitDefinition):
         self.module = bit_definition.module
+        self.cache_lifetime = bit_definition.cache_lifetime
+        self.bit_cache: dict[str, tuple[list[OOI], datetime]] = {}
 
-    def run(self, *args, **kwargs) -> list[OOI]:
+    def _purge(self) -> None:
+        now = datetime.now()
+        map(self.bit_cache.pop, [key for key, data in self.bit_cache.items() if now - data[1] <= self.cache_lifetime])
+
+    def _bit_cache_key(self, source: OOI, parameters: list[OOI], config: dict[str, JsonValue]) -> str:
+        try:
+            serialized_ooi = str(TypeAdapter(list[OOI]).dump_json(parameters))
+            return source.model_dump_json() + serialized_ooi + json.dumps(config)
+        except Exception:
+            return str(datetime.now())
+
+    def run(self, source: OOI, parameters: list[OOI], config: dict[str, JsonValue]) -> list[OOI]:
         module = import_module(self.module)
 
         if not hasattr(module, "run") or not isfunction(module.run):
@@ -29,7 +47,16 @@ class BitRunner:
             raise ModuleException(
                 f"Invalid run function return annotation, expected '{BIT_SIGNATURE.return_annotation}'"
             )
-        return list(module.run(*args, **kwargs))
+
+        if CACHE_BITS:
+            self._purge()
+            key = self._bit_cache_key(source, parameters, config)
+            if key not in self.bit_cache:
+                data = list(module.run(source, parameters, config=config))
+                self.bit_cache[key] = (data, datetime.now())
+            return self.bit_cache[key][0]
+        else:
+            return list(module.run(source, parameters, config=config))
 
     def __str__(self):
         return f"BitRunner {self.module}"
