+++ octopoes/octopoes/config/settings.py
@@ -89,3 +89,4 @@ DEFAULT_SEVERITY_FILTER = {severity for severity in RiskLevelSeverity}
 DEFAULT_LIMIT = 50
 DEFAULT_OFFSET = 0
 QUEUE_NAME_OCTOPOES: str = "octopoes"
+CACHE_BITS = True
